Added new cases to translated-up
translate-up now should handle the case where one change is "inside" the
bindings of another change, so their locations intersect. For instance,
in the case: (+ x (+ (exp (log y)) z)) -> (+ (+ x (exp (log y))) z)
-> (+ (+ x y) z), the first change operates on the whole
expression, (location: '()), and the second change operates on part
of the expression (location: '(cdr car cdr cdr car)). These two changes
are not orthgonal? since one takes place inside the other, but since the
second change is within the bindings of the first, we can still
translate the changes through each other. This functionality should be
added to translated-down in the near future. As in the last commit, this
functionality is mostly untested, since creating a test case is
currently pretty difficult.