Translate Up fully tested and working.
translate-up now works fully for the following use cases: If the two
changes to be swapped are orthogonal?, or one change is within the
bindings of the other, and that particular binding is one to one, as in
the variable in which the inner change is bound only appears once in
both the input and output patterns. This has finally been tested and can
be verfied to work.
This commit also includes the utility function is-inside?, which checks
whether location a is inside of location b in some manner (Note: this
does not check whether the change at location a is inside the bindings
of the change at location b, and a change can be inside another, and
still fail to translate through it.)
This commit also includes a refactoring of translate-up and the
beginnings of work on translate-down.