Rewrote Salmon Algorithm.
Originally I had the salmon algorithm (remove-red) moving all changes
upstream but preserving the changes behind them. This was because the
logic for moving the head upstream was very similar to the logic for
moving other changes upstream. Since I wanted to move the head until it
hit something it couldn't move past, then move that thing down
recursively, and retry if the thing was successfully moved, it made
sense to use the same logic for moving the head as moving other
changes. But since other changes needed to preserve the changes they
moved past, because those changes were still upstream from the head, I
had the move upstream logic always preserve the changes that they were
moving past, and just stripped non-green changes off the top at the
end. With this commit, the movement of the head upstream is delegated to
a seperate function, one which discards changes as the head moves past
them. This approach has two benefits: firstly, it allows us to move the
head upstream past changes even if we don't currently support moving
those changes down past the head, as in the case of the head having
non-one-to-one bindings and containing the other change. The second
benefit is that remove-red is now independent of green detection, and
will always properly eliminate changes that the head doesn't depend on,
even if you pass it a head that isn't green.