Added step 1 of simplification
simply removes instances of a function wrapped around it's inverse. I
maintained a seperate list of functions and their inverses. In the
future, this could also be accomplished by marking some of the rules as
"simplifying rules", rules like (log (exp x)) -> x, but that would
involve adding another field to the struct. This could also possibly
help with tighter rule integration in the future, if we chose to pursue
it. Anyway, this way works fine for now.