More bugfixes
I should really test, THEN commit. This is now tested pretty
thoroughly. Basically, it would fail on expressions like
'(* x (/ (- x x) (- x x))) because that simplified to
'(* x (/ 0 0)), and the (/ 0 0) got precomputed to nan. Now
precomputation checks for this, and returns the original expression if
the precomputed value is nan, because the simplification routine can
easily handle (/ 0 0) later on in the process, converting it to one.