Moved Some Alt Helper Functions to Alternative.rkt
While working on regime changes, I found that the helper functions
build-alt (which builds an alt from an initial alt and a list of
changes), alt-initial (which gets the original form of an alt), and
alt-changes (which gets a list of the changes that happened to an alt),
were pretty useful, so I added them to combine-alts.rkt. Now I've found
cases where I might want to use them elsewhere, so I moved them to
alternative.rkt, which seems like a more fitting place for them.