Can Now Combine Two Alts
Ultimately the goal is that you should be able to pass in all of your
alts, and an improvement function, and combine-alts will find the best
two to combine into a conditional alt. We don't have that yet. We do,
however, have combine-two-alts, so if you already know what alts you
want to combine, you just pass the index of the variable to branch
on (if your program is something like (lambda (x y) ...), and you want
to branch on x, that's an index of 0, 1 for y, etc.), and the two
alternatives, as well as an optional improvement function (which you'll
probably always use), and it'll combine the alts for you. Your
improvement function will have *points* dynamically scoped so that it
equals the points on which the alt you're currently improving is in
effect, so you can just do a normal improve and as long as you don't
override *points*, it'll improve on the right points.