Moved Makefile and Reports to Top-Level Directory
In order to maintain greater compatibility with the gh-pages branch, I
moved the makefile to the top level casio directory, so that the reports
will be generated there. Since this changes some relative directories,
it was necessary to make the proper changes to load-bench.rkt and
make-report.rkt as well.