Changed Reporting Of Error
Before, we just used the green-detection method of reporting error,
distilling error down to a single number for comparison by giving a
value to infinities. Now, we instead average the reasonable-error?
points in the Improvement column, and also report the number of positive
infinities as "Immeasurable Improvement", and the numbef of negative
infinities as "Immeasurable Regression."