Small Fix
Before, get-improvement-columns was summing over only the real elements
of the error difference, and then dividing by the length of the whole
errors difference. Now, we only divide by the length of the reals,
resulting in more accurate error reporting. Also in this commit: Added a
comment to explain a piece of seemingly useless code in
get-improvement-columns, and changed the annotated-diff
from (start-errors end-errors diff) to (diff start-errors end-errors).