Fixed Typo and Limited rsync
Don't know how it didn't casue a problem, but there was an $IDER instead
of $IDIR, so I fixed that. Also, rsync was copying across all the
reports when it copied the index, and that might not be what we want to
I limited the second rsync call to just the index files.