Made Red Highlighting More Picky
Previously any row with negative improvement was highlighted red. Due to some of the randomness inherent in our evaluation however, programs whose behaviour has not changed are usually given a score between -1 and 1 of improvement. Changed the red highlighting condition to be when improvement is less than -1 instead of when improvement is less than zero.