Made Epsilon For Binary Search Scaled
To find the proper splitpoint, we binary search between the points in
our pointslist which appear to be closest to what we want. Our binary
search stops when our low and hi are within epsilon of each other. While
before epsilon was a constant, now it is a fraction of the difference
between the start point of our search and it's end, otherwise points
with greater exponents would take a lot longer to find their results,
while points with very small exponents would have too course a search.