Simplified Some Logic
Initially I was setting the initial current region to null, and then
checking for null when adding the current region to the accumulator, but
that seemed messy. Now, we set the initial current region to the symbol
of the first item, and we don't have to do that check anymore.