Fixed Potential Bug On Inner Branches.
When we're deciding which option is the best, we evaluate the options on
a new set of points. But what if we're already in a branch, and we don't
have to worry about certain point intervals? Now, we have a dynamically
scoped predicate that allows us to filter out points that our branch
doesn't use so that we only worry about the right points.