Alt Filtration
We now export a function, plausible-alts, which can be given a bunch of
alts, and it returns the alts which could be useful for combination. It
is suggested that clients of combine-alts.rkt filter their alts with
plausible-alts before passing them to best-combination, for better
performance. In cases where you already know that your alts are all
plausible, this step may be skipped.