Removed Non-Star Functions
There's a couple functions that build scales, that are marked with a
start because they return two functions, the scale and it's inverse. I
had non-star versions of them for legacy reasons which basically just
called the star versions and threw out half the result, but I wasn't
using them and didn't want to worry about maintaining them when they
weren't used anywhere, so I took them out.