Added Depth and Tag Filtering To Debug Messages
The debug messages that were already in place still work as they did,
but now you can specify an optional keyword argument #:depth in the
debug command, setting how "deep" this message should be considered. You
can also use set-debug-level! to set the depth for which debug messages
from a particular location appear. For example, (set-debug-level!
'simplify 2) means that messages tagged as being from 'simplify with
depth of 2 or less will be displayed.