You can now turn off debug messages.
Added a check to make sure that when the max-depth level of a #:from is
non-positive, it never displays, instead of still displaying any debug messages
that have depth of zero or negative. This ensures that you can turn off messages
from a particular section by setting it to zero, and enforces #:depth's
of calls being positive.