Added Cycles/Hardness System.
The alt structure now contains a field called "cycles," and the change
structure has a cooresponding field called "hardness." Essentially, each
alternative starts with a cycles field of zero. Each time the main loop
decides to develop a particular alternative, it increments the cycles
field on all the other alternatives, and sets the "hardness" field of
the changes resulting from that development to the cycles of the
alternative it came from. This means that the cycles field of the
alternative is how many other alternatives have been considered since it
was created, and the hardness of a change indicates how many other paths
casio attempted to take before finding that change. This should allow us
to better understand which changes are easy for casio to find, and which
are hard.