Fixed Problem With Test Highlighting
Some of the indexes for columns switched around when I added the test and I forgot to change the cooresponding good/bad conditions, so if a test passed, the report code would interpret that as the test crashing, and highlight it red. I fixed the index, so now we've got proper highlighting in that area.