Changed find-match-loc Interface a Bit
Now, instead of just returning the location where a match was found on
success, it returns a values of the location and the bindings for the
match, so you don't have to recompute them later (although it would be
fairly trivial to do so). On failure returns a values of false and false.