Moved Error Computation Into avg-bits-error
There are a lot of error functions that do unclear things, so I added an
avg-bits-error function which, as the name suggests, calculates the
average error in bits across an error list. This takes care of logging
the errors, converting full error (nan, inf) to 64 bits of error, and such.