Reintroduced Regime Changes, Stopped From Spinning Out
With the new main loop, we keep making changes as long as we're making
green changes, and then we reduce our green threshold and see if that
helped, and keep reducing until we reach some minimum threshold. The
'fuel' is used to determine how many times we should step down before
reaching the minimum, and thus is the inverse of the size of the
step. The problem was, when we recursed into a regime, it's very easy on
the smaller number of points to do something trivial which over that
restricted region improves the average error bits by a fair
amount. Thus, keeping the thresholds the same, whenever we recursed on a
suffficiently small region, we would spin out, making stupid changes and
growing the size of our programs with log exps and sqr sqrts. Now, when
we recurse into smaller regions, we raise the thresholds proportional to
the region size change, so that we'll still only accept meaningful
changes. This appears to have stopped the spinning out.