Fixed Splitindex->Splitpoint Conversion.
Added a special case to the conversion that looks for splitindices with
the index of the length of the list (one past the allowed indices), and
converts it to +inf.0. This is because all points past the second to
last splitpoint should go to the alt specified by the last splitpoint,
and setting the "less than" point to +inf.0 makes those comparisons work out.