Added Better Special Case Handling
When we end up on a point range where the number of splitpoints we want
to make would split the points into less than the
minimum-region-size. As in, when min-region-size * num-splitpoints >
num-points. Now we're more careful to make sure to even in these cases
we save our results throughout the pipeline.