Made Stitching Differentiate Like Partitioning
So, in the regimes process, we have to play around a lot with
errors. When we take alts that were previously improved along one set of
points, and improve each along a subset of those points, we have to give
them only the errors and points for that subset, and then when we want
to get the errors of the whole combination at the end, we need to stitch
all the errors of the individual alts together to form the errors of our
combination. Well, it turned out that when we partitioned, we used a <=
comparison, and in that same place in stitching, we used a < comparison,
which was causing us to not stitch properly. That's fixed now.