Added map-2values helper function
I noticed that I was often writing functions that returned two values,
and then I would want to map those functions over lists and get two
lists, one for all the value1's that the function produced, and one for
all the value2's that it produced. So I wrote a function that abstracts
that away.