Fixed Issue With (sqrt (sqr x))
Since (sqrt (sqr x)) is equivilent to (abs x) and not just x, things
surrounding it get tricky. Normally, if you have an expression of the
form (expt (expt x 3) 2), then this is the variable x raised to the
3*2=6th power. With the sqrt-sqr case though, multiplying the powers
yields one, making it seem like you could cancel it with a x, even
though you can't. Special casing is required.