Added New Coerece-Inverses Phase
This new phase turns expressions that are left of the form (/ a) into (/
1 a), because then we can recongnize more transformations for the rest
of casio. This happens seperate from the normalize step since many
expressions of the form (/ a) will be turned into other things during
normalize. Both this and the normalize phase are now under the umbrella
of 'cleanup'