Better pareto alt minimization.
Minimization removes alts from the alt-table when there are alts
which are not best at any point. We must choose an alt to remove.
Previously, we compared the alt costs and errors; now we default
to removing alts that haven't been explored yet (try to saturate)
and alts that are tied on few points (less likely to help).