Added System So That Iterations Are More Predictable
When a node is first created, it is marked as not real, so that no other
rule application that iteration can match against it. When the iteration
is over, all nodes in the tree are marked as real again. Not entirely
foolproof, since two nodes can still be merged that allow a match if
they were both already existing but in different places, but works
pretty well.