Fixed bug, cleaned up merge function
Fixed a bug where a non-leader was coming out of the merge function,
causing a hash lookup to fail. Turned out to be due to a merge of two
enodes sometimes cascading into a merge of more enodes into the same
pack, because of cycles in the egraph. Now we just update to the leader
after the cascade of merges. Also added a bunch more comments and
cleaned up some code, especially in the merge function.