Add support for fused multiply-add
This patch changes a few implementations of operators to use the versions
from libm via FFI.  These implementations should more representative of
behavior for users on various platforms.

This also fixes a bug in simplify precompute which was not translating from
operator names to their actual implementations when evaluating constant
expressions.