Fix null points bug
The issue here was rather complex. At the first layer, we were
creating conditionals "l < EXPR < u" where no sampled point actually
evaluated that expression between "l" and "u". The reason for this is
that were evaluating the expression in arbitrary
precision (sometimes). But upon fixing that, we were getting a lot
more NaNs, so we also needed to fix up sorting to use a total sort
order (where NaN is less than anything).