Early exit speeds up rewriting on niche examples
In some cases when we're doing the recursive rewrite, we are trying to
match a pattern like (f A B C) against some expression (g D E F). To
do that, we first try rewriting D to match A, E to match B, etc. Each
of these calls returns a list of options, and then we cartesian
product them together and combine them to get a way to rewrite the
whole expression.

Now, in some cases, there might be no ways to rewrite D to match A.
This patch makes the matcher exit early, so that in this case, we'd
stop immediately and return the empty list from the cartesian product.

In the few cases where we've seen rewriting be slow, this commit fixes it.