Remove `Runner` invalidation in favor of reconstructing the `Box`
The `Box` has special compiler magic that lets us move ownership of
inner data on `T`. This means we don't have to use `take` to put a
`None` on the heap while we use it. In most places, were we don't move
ownership of the `EGraph`, we can use `ManuallyDrop` to prevent the
destructor from running, but I had to use `mem::forget` in two places
where this didn't work.