Add Resyntax Autofixer workflow
As discussed at RacketCon, this pull request adds the Resyntax Autofixer workflow to this repository. Once a week, this workflow will run Resyntax on the repository. Resyntax selects a small handful of suggested fixes and puts them into a new pull request. Configurable limits are placed on the number of fixes applied, how many files are touched, and the size of the pull request. I've set those limits quite low to start but feel free to change them to whatever suits your process.

You'll also need to add a secret named `RESYNTAX_APP_PRIVATE_KEY` to the repository. I can send you a key out-of-band.