Use the same cost model throughout `cost-opportunity`
The `cost-opportunity` heuristic uses the cost model in two ways.
First, the cost model is passed to the egraph to find the simplest
form of any given expression. And then, the cost model is used again
to compute cost opportunity.

The bug is that we were using different cost models for these two
phases; default cost model for building the egraph and platform cost
model for computing cost opportunity. This lead to weird stuff like
the cost opportunity of a node being negative.

The fix is to use the same cost model for both phases; the code does
that though note that it is a little ugly in that it duplicates some
existing code.